# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ambulance & Emergency Warehouse Management System API — Damascus
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List items with optional filters
      parameters:
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: belowMin
          in: query
          schema: { type: boolean }
        - name: nearExpiry
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 5000 }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, currentStock, minStock, expiryDate, createdAt]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Paginated list of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
    put:
      operationId: updateItem
      tags: [items]
      summary: Update item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Deactivate item (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deactivated
  /items/history:
    get:
      operationId: getItemHistory
      tags: [items]
      summary: Get an item card with its paginated chronological movement history
      parameters:
        - name: itemId
          in: query
          required: true
          schema: { type: integer }
        - name: type
          in: query
          schema:
            type: string
            enum: [in, init, out, adjust, custody_out, custody_return, damage, central_return]
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: document
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: Item details, batches, and chronological movement history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemHistoryResponse"
        "400":
          description: Invalid filter
        "404":
          description: Item not found
...
  /items/fefo-preview:
    get:
      operationId: getItemFefoPreview
      tags: [items]
      summary: Preview FEFO allocations for a consumable item
      parameters:
        - name: itemId
          in: query
          required: true
          schema: { type: integer }
        - name: quantity
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: FEFO allocation preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FefoPreview"
        "400":
          description: Invalid quantity or non-consumable item
        "404":
          description: Item not found
...
  /transactions/out:
    post:
      operationId: createOutTransaction
      tags: [transactions]
      summary: Record an outbound transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Item:
      type: object
      required: [id, name, unit, itemType, currentStock, minStock, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: ["string", "null"]
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        itemType:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minStock:
          type: integer
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ItemInput:
      type: object
      required: [name, unit, itemType]
      properties:
        code:
          type: ["string", "null"]
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        itemType:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minStock:
          type: integer
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ItemUpdate:
      type: object
      properties:
        code:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        itemType:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        minStock:
          type: ["integer", "null"]
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    ItemListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ItemHistoryResponse:
      type: object
      required: [item, batches, movements, total, page, limit]
      properties:
        item:
          $ref: "#/components/schemas/Item"
        batches:
          type: array
          items:
            $ref: "#/components/schemas/ItemHistoryBatch"
        movements:
          type: array
          items:
            $ref: "#/components/schemas/ItemHistoryMovement"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Transaction:
      type: object
      required: [id, type, itemType, documentNumber, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out, init, adjust, custody_out, custody_return, damage, central_return]
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        itemName:
          type: ["string", "null"]
        itemUnit:
          type: ["string", "null"]
        equipmentId:
          type: ["integer", "null"]
        equipmentName:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientName:
          type: ["string", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        exitReason:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        supplySource:
          type: ["string", "null"]
          enum: [central_warehouses, null]
        documentDate:
          type: ["string", "null"]
        deliveryNoteNumber:
          type: ["string", "null"]
        deliveryNoteDate:
          type: ["string", "null"]
        internalDeliveryNoteNumber:
          type: ["string", "null"]
        internalDeliveryNoteDate:
          type: ["string", "null"]
        deliveryDestination:
          type: ["string", "null"]
          enum: [administrative_building, ambulance_point, null]
        custodyHolderName:
          type: ["string", "null"]
        custodyNoteNumber:
          type: ["string", "null"]
        custodyDate:
          type: ["string", "null"]
        custodyLocation:
          type: ["string", "null"]
        returnCondition:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        documentNumber:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    FefoPreview:
      type: object
      required: [itemId, itemStock, requestedQuantity, availableQuantity, canFulfill, allocations, expiredBatches]
      properties:
        itemId:
          type: integer
        itemStock:
          type: integer
        requestedQuantity:
          type: integer
        availableQuantity:
          type: integer
        canFulfill:
          type: boolean
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/FefoAllocation"
        expiredBatches:
          type: array
          items:
            $ref: "#/components/schemas/FefoExpiredBatch"
...
    OutTransactionInput:
      type: object
      required: [itemType, itemId, quantity, recipientId, exitReasonId, internalDeliveryNoteNumber, internalDeliveryNoteDate, deliveryDestination]
      properties:
        itemType:
          type: string
          enum: [item]
        itemId:
          type: integer
        equipmentId:
          type: ["integer", "null"]
        quantity:
          type: integer
          minimum: 1
        recipientId:
          type: integer
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        internalDeliveryNoteNumber:
          type: string
          minLength: 1
        internalDeliveryNoteDate:
          type: string
          format: date
        deliveryDestination:
          type: string
          enum: [administrative_building, ambulance_point]
        notes:
          type: ["string", "null"]
...